Fix shape and length reporting of gridded interfaces for Python 3.6 on Win 64.

+++ holoviews/core/data/xarray.py
@@ -50,7 +50,7 @@ class XArrayInterface(GridInterface):
         if gridded:
             return shape
         else:
-            return (np.product(shape), len(dataset.dimensions()))
+            return (np.product(shape,dtype=np.intp), len(dataset.dimensions()))
 
 
     @classmethod
